struct ndicapi defined in ndicapi.h instead of ndicapi.cxx
